🔨 refactor: templates methods of measurement module now pass to common module
